Revert "Added W16_lab05.md"
