+++ .gitignore
@@ -1,7 +1,6 @@
 .DS_Store
 
 *.class
-savedLists
 
 # Package Files #
 *.jar
