+++ build.xml
@@ -3,7 +3,7 @@
   <!-- build.xml for todo
        Note: for this lab, you should not need to edit this file -->
 
-   <property name="projectPackage" value="edu.ucsb.cs56.projects.utilities.todo" />
+   <property name="projectPackage" value="edu.ucsb.cs56.projects.utilties.todo" />
    <property environment="env"/> <!-- load the environment variables --> 
    <property name="webRoot" value="${env.HOME}/public_html/cs56" />
    <property name="webBaseURL" value="http://www.cs.ucsb.edu/~khalid/cs56" />
